spellcheck/scripts/batch/requirements.txt
vllm==0.5.4
google-cloud-storage==2.18.0